+++ CHANGELOG.rst
@@ -19,6 +19,10 @@ v34.5.0 (unreleased)
   SBOMs that contains properties with no values.
   https://github.com/nexB/scancode.io/issues/1185
 
+- Add a new `CollectTreeSitterSymbolsAndStrings` pipeline (addon) for collecting source
+  symbol and string using tree-sitter.
+  https://github.com/nexB/scancode.io/pull/1181
+
 v34.4.0 (2024-04-22)
 --------------------